apps/ocr/requirements.txt
# OCR Engine
rapidocr>=3.0.0
onnxruntime>=1.16.0

# MRZ parsing (ICAO 9303 standard)
mrz>=0.6.0

# Image processing
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Typing
pydantic>=2.0.0
